Initial creation of the hugo-based website
This commit imports a stripped down version of the website
inherited from the submariner project.

It provides basic Makefile, and github scripts for publishing,
setting up a GIT_DEPLOY_KEY secret (private ssh key) for this project
and and the public counterside on ovn-org.github.io will be necessary
for the "publish/deploy" part to work properly.