Add Release content for OVN
For the time being, this is being presented as static markdown. This at
least gets us an actual "Releases" page, and it isn't too difficult to
amend by hand as we create new releases.

A future improvement would be to generate the releases page with a script
so that updating the releases would be even easier in the future.

Signed-off-by: Mark Michelson <mmichels@redhat.com>