site-wide: Generate site content with a maintenance script.
When OVN 23.09 is released, we also will begin making more frequent
point-releases of all other supported branches. Updating the website
by hand for each release would be time-consuming and error-prone.

This change uses a website update script to generate the website content.
The script can be used for generating content for a new release or for
updating content for previous releases, since the script re-generates
content from scratch instead of actually updating the content.