ci: Fix automation for dist-docs
The automation for dist-docs was broken because of the PR action.
Fix the automation, align it more with current state of OVN ci
and change the branches that should be updated according to current
supported branches.

The updated action takes less time to complete (~3 minutes per job)
and the steps should be clearer.

Signed-off-by: Ales Musil <amusil@redhat.com>